Css experiments

+++ Src/Engine.cs
@@ -51,9 +51,18 @@ namespace Csml {
             //typeof(T).GetFields()
         }
 
-        public static void Generate<R>(Context context) {
-            if (Directory.Exists(context.OutputRootDirectory))
-                Directory.Delete(context.OutputRootDirectory, true);            
+        public static void Generate<R>(Context context, bool clean = true) {
+            if (clean) {
+                if (Directory.Exists(context.OutputRootDirectory))
+                    Directory.Delete(context.OutputRootDirectory, true);
+            }
+
+            foreach (var i in context.AssetsToCopy) {
+                var dest = Path.Combine(context.OutputRootDirectory, Context.GetContentRelativePath(i, context.SourceRootDirectory));
+                Utils.CreateDirectories(Path.GetDirectoryName(dest));
+                if (!File.Exists(dest) | context.ForceRebuildAssets)
+                    File.Copy(i, dest, true);
+            }
 
             var fields = typeof(R).GetNestedTypes().SelectMany(x => x.GetFields(BindingFlags.Static | BindingFlags.Public));
             foreach (var f in fields) {
